Split care service into smaller hooks
