+++ src/App.vue
@@ -608,734 +608,60 @@
 </template>
 
 <script setup>
-import { ref, computed, watch, onMounted } from 'vue'
-import { useQuasar } from 'quasar'
-import { gsap } from 'gsap'
-
-// 導入 Chart.js 相關組件
-import {
-  Chart as ChartJS,
-  Title,
-  Tooltip,
-  Legend,
-  ArcElement,
-  CategoryScale,
-  LinearScale
-} from 'chart.js'
 import { Pie } from 'vue-chartjs'
+import useCareService from './composables/useCareService'
 
-// 註冊 Chart.js 組件
-ChartJS.register(CategoryScale, LinearScale, ArcElement, Title, Tooltip, Legend)
-
-const $q = useQuasar()
-
-// 粒子加載完成回調
-const particlesLoaded = async (container) => {
-  console.log('粒子容器已加載', container)
-}
-
-// 服務項目數據
-const careItems = ref([
-  // 鐘點制項目 - 基本服務
-  { code: 'HR01', name: '基本照護(必選)', price: 200, category: '鐘點制', subCategory: '基本服務', selectedByDefault: true, icon: 'favorite', color: 'deep-orange' },
-  
-  // 鐘點制項目 - 醫療照護
-  { code: 'HR02', name: '抽痰需求', price: 50, category: '鐘點制', subCategory: '醫療照護', icon: 'medical_services', color: 'light-blue' },
-  { code: 'HR03', name: '鼻胃管或其他管灌', price: 60, category: '鐘點制', subCategory: '醫療照護', icon: 'medical_services', color: 'teal' },
-  { code: 'HR04', name: '導尿管或其它引流管', price: 50, category: '鐘點制', subCategory: '醫療照護', icon: 'medical_services', color: 'cyan' },
-  { code: 'HR05', name: '氣切or氣管內管', price: 60, category: '鐘點制', subCategory: '醫療照護', icon: 'medical_services', color: 'indigo' },
-  
-  // 鐘點制項目 - 特殊需求
-  { code: 'HR06', name: '體重>50公斤並需協助上下床', price: 70, category: '鐘點制', subCategory: '特殊需求', icon: 'bed', color: 'deep-purple' },
-  { code: 'HR07', name: '意識不清或無法配合', price: 70, category: '鐘點制', subCategory: '特殊需求', icon: 'psychology', color: 'purple' },
-  { code: 'HR08', name: '有任何傳染病或需要隔離病人', price: 80, category: '鐘點制', subCategory: '特殊需求', icon: 'coronavirus', color: 'red' },
-  
-  // 鐘點制項目 - 時段加價
-  { code: 'HR09', name: '夜間:2200-0800', price: 20, category: '鐘點制', subCategory: '時段加價', icon: 'bedtime', color: 'blue-grey' },
-  { code: 'HR10', name: '加價急徵', price: 30, category: '鐘點制', subCategory: '時段加價', icon: 'priority_high', color: 'amber' },
-  { code: 'HR11', name: '二次加價急徵', price: 30, category: '鐘點制', subCategory: '時段加價', icon: 'bolt', color: 'orange' },
-  
-  // 包班制項目 - 基本班次
-  { code: 'SH01', name: '12小時/班', price: 3000, category: '包班制', subCategory: '基本班次', icon: 'schedule', color: 'primary' },
-  { code: 'SH02', name: '24小時/班', price: 5500, category: '包班制', subCategory: '基本班次', icon: 'access_time_filled', color: 'secondary' },
-  
-  // 包班制項目 - 特殊需求
-  { code: 'SH03', name: '體重>50公斤並需協助上下床', price: 200, category: '包班制', subCategory: '特殊需求', icon: 'bed', color: 'deep-purple' },
-  { code: 'SH04', name: '意識不清或無法配合', price: 200, category: '包班制', subCategory: '特殊需求', icon: 'psychology', color: 'purple' },
-  { code: 'SH05', name: '有任何傳染病或需要隔離病人', price: 300, category: '包班制', subCategory: '特殊需求', icon: 'coronavirus', color: 'red' }
-])
-
-// 狀態變數
-const selectedHourlyItems = ref(careItems.value.filter(item => item.selectedByDefault && item.category === '鐘點制'))
-const selectedShiftItems = ref([])
-const selectedCategory = ref('鐘點制')
-const selectedShiftType = ref('SH01') // 預設選擇12小時班
-const searchText = ref('')
-const showHelp = ref(false)
-const previousTotalCost = ref(0)
-const totalCostElement = ref(null)
-const particleContainer = ref(null)
-const tipAmount = ref(0) // 小費金額
-const hourCount = ref(1) // 鐘點制小時數
-const dayCount = ref(1) // 鐘點制天數
-const shiftDayCount = ref(1) // 包班制天數
-
-// 收合/展開狀態
-const showFilters = ref(false)
-const showCalculator = ref(false)
-const showShiftCalculator = ref(false)
-const showAdditionalItems = ref(true)
-
-const isNightShift = ref(false) // 是否為夜班時段
-const isUrgent = ref(false) // 是否為加價急徵
-const isDoubleUrgent = ref(false) // 是否為二次加價急徵
-
-
-// 篩選系統狀態
-const activeFilters = ref({
-  price: 0,
-  subCategory: null
-})
-
-// 判斷是否有任何篩選條件被啟用
-const isAnyFilterActive = computed(() => {
-  return activeFilters.value.price > 0 || 
-         activeFilters.value.subCategory !== null
-})
-
-// 粒子效果配置
-const particlesOptions = {
-  background: {
-    color: {
-      value: "transparent"
-    }
-  },
-  fpsLimit: 60,
-  particles: {
-    color: {
-      value: ["#4A90E2", "#50C8B4", "#89253e"]
-    },
-    links: {
-      color: "#c8c8c8",
-      distance: 150,
-      enable: true,
-      opacity: 0.2,
-      width: 1
-    },
-    move: {
-      direction: "none",
-      enable: true,
-      outModes: {
-        default: "bounce"
-      },
-      random: true,
-      speed: 1,
-      straight: false
-    },
-    number: {
-      density: {
-        enable: true,
-        area: 800
-      },
-      value: 30
-    },
-    opacity: {
-      value: 0.3
-    },
-    shape: {
-      type: "circle"
-    },
-    size: {
-      value: { min: 1, max: 3 }
-    }
-  },
-  detectRetina: true
-}
-
-// 計算屬性
-const selectedItems = computed(() => 
-  selectedCategory.value === '鐘點制' ? selectedHourlyItems.value : selectedShiftItems.value
-)
-
-// 計算鐘點制每小時費用是否達到上限
-const hourlyRate = computed(() => {
-  if (selectedCategory.value !== '鐘點制') return 0
-  
-  return selectedHourlyItems.value.reduce((sum, item) => {
-    return item.subCategory === '時段加價' ? sum : sum + item.price
-  }, 0)
-})
-
-// 判斷是否達到鐘點制上限 (不含小費)
-const hourlyRateReachedLimit = computed(() => {
-  return hourlyRate.value >= 500
-})
-
-// 移除500元限制，永遠返回false
-function wouldExceedLimit(item) {
-  return false;
-}
-
-// 鐘點制費用計算（考慮小時數和天數）
-const hourlyTotalWithTime = computed(() => {
-  if (selectedCategory.value !== '鐘點制') return 0
-  
-  // 基本費用（每項服務的價格）
-  return selectedHourlyItems.value.reduce((sum, item) => {
-    // 時段加價項目不要乘以小時數
-    if (item.subCategory === '時段加價') {
-      return sum + (item.price * dayCount.value)
-    }
-    return sum + (item.price * hourCount.value * dayCount.value)
-  }, 0)
-})
-
-
-
-// 可用的子類別列表
-const availableSubCategories = computed(() => {
-  // 根據當前選擇的主類別，獲取可用的子類別
-  const subCategories = careItems.value
-    .filter(item => item.category === selectedCategory.value && (item.subCategory === '醫療照護' || item.subCategory === '特殊需求'))
-    .map(item => item.subCategory)
-  
-  // 去除重複項
-  return [...new Set(subCategories)]
-})
-
-// 班次選項
-const shiftTypeItems = [
-  { code: 'SH01', name: '12小時/班', price: 3000, category: '包班制', subCategory: '基本班次', icon: 'schedule', color: 'primary' },
-  { code: 'SH02', name: '24小時/班', price: 5500, category: '包班制', subCategory: '基本班次', icon: 'access_time_filled', color: 'secondary' }
-];
-
-const filteredItems = computed(() => {
-  return careItems.value.filter(item => {
-    const matchCategory = item.category === selectedCategory.value
-    let matchSearch = true
-    let matchPrice = true
-    let matchSubCategory = true
-    
-    // 如果是包班制，只顯示特殊需求項目，不顯示基本班次項目
-    if (selectedCategory.value === '包班制' && item.subCategory === '基本班次') {
-      return false;
-    }
-    
-    // 處理搜尋邏輯
-    if (searchText.value) {
-      matchSearch = item.name.toLowerCase().includes(searchText.value.toLowerCase())
-    }
-    
-    // 處理價格篩選
-    if (activeFilters.value.price > 0) {
-      matchPrice = item.price > activeFilters.value.price
-    }
-    
-    // 處理子類別篩選
-    if (activeFilters.value.subCategory) {
-      matchSubCategory = item.subCategory === activeFilters.value.subCategory
-    }
-    
-    return matchCategory && matchSearch && matchPrice && matchSubCategory
-  })
-})
-
-// 包班制費用計算（考慮天數）
-const shiftTotalWithDays = computed(() => {
-  if (selectedCategory.value !== '包班制') return 0
-  
-  // 基本班次費用
-  let basePrice = 0;
-  if (selectedShiftType.value === 'SH01') {
-    basePrice = 3000; // 12小時班費用
-  } else if (selectedShiftType.value === 'SH02') {
-    basePrice = 5500; // 24小時班費用
-  }
-  
-  // 包班制費用 = 班次基本費用 * 天數 + 附加服務費用 * 天數
-  const additionalServices = selectedShiftItems.value.reduce((sum, item) => sum + item.price, 0);
-  return (basePrice + additionalServices) * shiftDayCount.value;
-})
-
-const totalCost = computed(() => {
-  if (selectedCategory.value === '鐘點制') {
-    return hourlyTotalWithTime.value
-  } else {
-    return shiftTotalWithDays.value
-  }
-})
-
-// Chart.js 圖表數據
-const chartData = computed(() => {
-  if (selectedItems.value.length === 0) {
-    return {
-      labels: [],
-      datasets: []
-    }
-  }
-  
-  // 按子類別分組計算費用
-  const subCategoryCosts = {}
-  let totalCostBySubCategory = 0
-  
-  selectedItems.value.forEach(item => {
-    if (!subCategoryCosts[item.subCategory]) {
-      subCategoryCosts[item.subCategory] = 0
-    }
-    subCategoryCosts[item.subCategory] += item.price
-    totalCostBySubCategory += item.price
-  })
-  
-  // 轉換為圖表數據格式
-  const labels = Object.keys(subCategoryCosts)
-  const data = labels.map(label => subCategoryCosts[label])
-  
-  // 為不同子類別設置不同顏色
-  const backgroundColors = [
-    '#4A90E2', // 藍色
-    '#50C8B4', // 綠色
-    '#F5A623', // 橙色
-    '#D0021B', // 紅色
-    '#9013FE', // 紫色
-    '#BD10E0', // 粉紅色
-    '#7ED321'  // 淺綠色
-  ]
-  
-  return {
-    labels: labels,
-    datasets: [
-      {
-        data: data,
-        backgroundColor: backgroundColors.slice(0, labels.length),
-        borderWidth: 0
-      }
-    ]
-  }
-})
-
-// Chart.js 圖表選項
-const chartOptions = {
-  responsive: true,
-  maintainAspectRatio: false,
-  plugins: {
-    legend: {
-      position: 'bottom',
-      labels: {
-        boxWidth: 12,
-        padding: 15,
-        font: {
-          size: 12
-        }
-      }
-    },
-    tooltip: {
-      callbacks: {
-        label: function(context) {
-          const label = context.label || '';
-          const value = context.raw || 0;
-          const percentage = ((value / totalCost.value) * 100).toFixed(1);
-          return `${label}: ${formatCurrency(value)} 元 (${percentage}%)`;
-        }
-      }
-    }
-  }
-}
-
-// 格式化金額函數
-function formatCurrency(value) {
-  return value.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
-}
-
-// 方法
-function toggleItem(item) {
-  const targetArray = selectedCategory.value === '鐘點制' ? selectedHourlyItems : selectedShiftItems
-  const index = targetArray.value.findIndex(i => i.code === item.code)
-  
-  // 如果是鐘點制且要添加非時段加價項目，且選擇後會超過上限，則不允許添加
-  if (selectedCategory.value === '鐘點制' && 
-      index === -1 && 
-      wouldExceedLimit(item)) {
-    // 顯示提示
-    $q.notify({
-      message: '鐘點制服務項目每小時不能超過500元上限',
-      color: 'warning',
-      icon: 'warning',
-      position: 'top',
-      timeout: 2000
-    })
-    return
-  }
-  
-  // 取得點擊位置，用於粒子效果
-  const clickEvent = window.event
-  if (clickEvent) {
-    createParticleExplosion(clickEvent.clientX, clickEvent.clientY, item.color || 'primary')
-  }
-  
-  if (index === -1) {
-    // 添加項目時的動畫效果
-    targetArray.value.push(item)
-    
-    // 使用 GSAP 對項目元素進行動畫
-    const itemElement = document.querySelector(`[data-code="${item.code}"]`)
-    if (itemElement) {
-      gsap.fromTo(itemElement, 
-        { scale: 1, backgroundColor: 'rgba(76, 175, 80, 0.2)' },
-        { scale: 1, backgroundColor: 'rgba(25, 118, 210, 0.08)', duration: 0.5, ease: 'power1.out' }
-      )
-    }
-  } else if (item.code !== 'HR01') {
-    // 移除非必選項目
-    targetArray.value.splice(index, 1)
-    
-    // 使用 GSAP 對項目元素進行動畫
-    const itemElement = document.querySelector(`[data-code="${item.code}"]`)
-    if (itemElement) {
-      gsap.fromTo(itemElement, 
-        { scale: 1, backgroundColor: 'rgba(255, 87, 34, 0.2)' },
-        { scale: 1, backgroundColor: 'transparent', duration: 0.5, ease: 'power1.out' }
-      )
-    }
-  }
-}
-
-// 創建粒子爆炸效果
-function createParticleExplosion(x, y, color) {
-  // 創建容器
-  const container = document.createElement('div')
-  container.style.position = 'fixed'
-  container.style.left = '0'
-  container.style.top = '0'
-  container.style.width = '100%'
-  container.style.height = '100%'
-  container.style.pointerEvents = 'none'
-  container.style.zIndex = '9999'
-  document.body.appendChild(container)
-  
-  // 創建粒子
-  const particleCount = 20
-  const particles = []
-  
-  for (let i = 0; i < particleCount; i++) {
-    const particle = document.createElement('div')
-    particle.style.position = 'absolute'
-    particle.style.width = `${Math.random() * 6 + 2}px`
-    particle.style.height = particle.style.width
-    particle.style.backgroundColor = color === 'primary' ? '#1976d2' : 
-                                    color === 'secondary' ? '#26a69a' : 
-                                    color === 'deep-orange' ? '#ff5722' : 
-                                    `#${Math.floor(Math.random() * 16777215).toString(16)}`
-    particle.style.borderRadius = '50%'
-    particle.style.left = `${x}px`
-    particle.style.top = `${y}px`
-    container.appendChild(particle)
-    particles.push(particle)
-    
-    // 使用 GSAP 動畫
-    const angle = Math.random() * Math.PI * 2
-    const speed = Math.random() * 100 + 50
-    const vx = Math.cos(angle) * speed
-    const vy = Math.sin(angle) * speed
-    
-    gsap.to(particle, {
-      x: vx,
-      y: vy,
-      opacity: 0,
-      duration: 0.6 + Math.random() * 0.4,
-      ease: 'power1.out',
-      onComplete: () => {
-        container.removeChild(particle)
-        // 清理容器
-        if (container.childElementCount === 0) {
-          document.body.removeChild(container)
-        }
-      }
-    })
-  }
-}
-
-function isSelected(item) {
-  return selectedItems.value.some(i => i.code === item.code)
-}
-
-// 選擇班次類型
-function selectShiftType(item) {
-  selectedShiftType.value = item.code;
-}
-
-function resetSelections() {
-  // 重置為初始狀態，保留必選項目
-  selectedHourlyItems.value = careItems.value.filter(item => item.selectedByDefault && item.category === '鐘點制')
-  selectedShiftItems.value = []
-  selectedShiftType.value = 'SH01' // 重置為12小時班
-  searchText.value = ''
-  
-  // 重置所有篩選條件
-  resetAllFilters()
-  
-  // 重置加價選項
-  isNightShift.value = false
-  isUrgent.value = false
-  isDoubleUrgent.value = false
-  
-  // 重置天數和小時數
-  dayCount.value = 1
-  hourCount.value = 1
-  shiftDayCount.value = 1
-}
-
-// 統一的篩選應用函數
-function applyFilter(filterType, value) {
-  // 清空搜尋文字，避免衝突
-  if (filterType !== 'search') {
-    searchText.value = ''
-  }
-  
-  // 如果點擊的是當前已選中的篩選條件，則取消該篩選
-  if (activeFilters.value[filterType] === value) {
-    activeFilters.value[filterType] = filterType === 'price' ? 0 : null
-    if (filterType === 'mandatory') {
-      activeFilters.value.mandatory = false
-    }
-  } else {
-    // 否則應用新的篩選條件
-    activeFilters.value[filterType] = value
-  }
-}
-
-// 重置所有篩選條件
-function resetAllFilters() {
-  activeFilters.value = {
-    price: 0,
-    subCategory: null
-  }
-  
-  // 顯示重置篩選通知
-  if (isAnyFilterActive.value) {
-    $q.notify({
-      message: '已清除所有篩選條件',
-      color: 'secondary',
-      icon: 'filter_alt_off',
-      position: 'top',
-      timeout: 1500
-    })
-  }
-}
-
-// 處理加價服務的切換
-function toggleAdditionalService(code) {
-  const item = careItems.value.find(item => item.code === code)
-  if (!item) return
-  
-  const index = selectedHourlyItems.value.findIndex(i => i.code === code)
-  
-  // 根據 toggle 狀態添加或移除服務
-  if (code === 'HR09') { // 夜間時段
-    if (isNightShift.value && index === -1) {
-      selectedHourlyItems.value.push(item)
-    } else if (!isNightShift.value && index !== -1) {
-      selectedHourlyItems.value.splice(index, 1)
-    }
-  } else if (code === 'HR10') { // 加價急徵
-    if (isUrgent.value && index === -1) {
-      selectedHourlyItems.value.push(item)
-    } else if (!isUrgent.value && index !== -1) {
-      selectedHourlyItems.value.splice(index, 1)
-    }
-  } else if (code === 'HR11') { // 二次加價急徵
-    if (isDoubleUrgent.value && index === -1) {
-      selectedHourlyItems.value.push(item)
-    } else if (!isDoubleUrgent.value && index !== -1) {
-      selectedHourlyItems.value.splice(index, 1)
-    }
-  }
-}
-
-function calculatePreviewCost(item) {
-  // 計算預覽費用
-  if (isSelected(item)) {
-    // 如果已選中，則顯示移除後的總費用
-    if (item.code === 'HR01') {
-      // 基本照護不可移除
-      return totalCost.value
-    }
-    
-    // 鐘點制需要考慮小時數
-    if (selectedCategory.value === '鐘點制' && item.subCategory !== '時段加價') {
-      return totalCost.value - (item.price * hourCount.value)
-    } else {
-      return totalCost.value - item.price
-    }
-  } else {
-    // 如果未選中，則顯示添加後的總費用
-    // 鐘點制需要考慮小時數
-    if (selectedCategory.value === '鐘點制' && item.subCategory !== '時段加價') {
-      return totalCost.value + (item.price * hourCount.value)
-    } else {
-      return totalCost.value + item.price
-    }
-  }
-}
-
-function getItemIcon(item) {
-  // 返回項目的圖標，如果沒有定義則使用默認圖標
-  return item.icon || 'help_outline'
-}
-
-function getItemColor(item) {
-  // 返回項目的顏色，如果沒有定義則使用默認顏色
-  return item.color || 'primary'
-}
-
-
-
-
-
-// 監聽總費用變化，添加動畫效果
-watch(totalCost, (newVal, oldVal) => {
-  previousTotalCost.value = oldVal
-  
-  // 使用 GSAP 實現數字變化動畫
-  if (totalCostElement.value) {
-    // 先停止任何正在進行的動畫
-    gsap.killTweensOf(totalCostElement.value)
-    
-    // 如果是增加費用，使用綠色動畫；如果是減少費用，使用紅色動畫
-    const color = newVal > oldVal ? '#4caf50' : newVal < oldVal ? '#f44336' : 'inherit'
-    
-    // 先變色
-    gsap.to(totalCostElement.value, {
-      color: color,
-      duration: 0.3,
-      ease: 'power2.inOut',
-      onComplete: () => {
-        // 然後返回原來的預設顏色
-        gsap.to(totalCostElement.value, {
-          color: 'inherit',
-          duration: 0.5,
-          delay: 0.5,
-          ease: 'power2.inOut'
-        })
-      }
-    })
-    
-    // 對數字進行彈跳動畫
-    gsap.fromTo(totalCostElement.value, 
-      { scale: 1 },
-      { scale: 1.2, duration: 0.2, ease: 'power2.out', yoyo: true, repeat: 1 }
-    )
-    
-    // 如果費用變化較大，則創建浮動數字效果
-    if (Math.abs(newVal - oldVal) > 100) {
-      const floatingNumber = document.createElement('div')
-      floatingNumber.textContent = newVal > oldVal ? `+${newVal - oldVal}` : `${newVal - oldVal}`
-      floatingNumber.style.position = 'fixed'
-      floatingNumber.style.color = newVal > oldVal ? '#4caf50' : '#f44336'
-      floatingNumber.style.fontSize = '1.5rem'
-      floatingNumber.style.fontWeight = 'bold'
-      floatingNumber.style.zIndex = '9999'
-      floatingNumber.style.pointerEvents = 'none'
-      
-      // 將浮動數字放在總費用附近
-      const rect = totalCostElement.value.getBoundingClientRect()
-      floatingNumber.style.left = `${rect.right + 10}px`
-      floatingNumber.style.top = `${rect.top}px`
-      
-      document.body.appendChild(floatingNumber)
-      
-      // 使用 GSAP 對浮動數字進行動畫
-      gsap.to(floatingNumber, {
-        y: -50,
-        opacity: 0,
-        duration: 1.5,
-        ease: 'power2.out',
-        onComplete: () => {
-          document.body.removeChild(floatingNumber)
-        }
-      })
-    }
-  }
-})
-
-
-
-// 鐘點制即時費用計算器
-function calculateHourlyCost() {
-  // 取得小時數輸入值
-  const hourCount = hourCountInput.value
-  
-  // 取得時段選擇值
-  const timePeriod = timePeriodSelect.value
-  
-  // 計算即時費用
-  const hourlyCost = calculateHourlyCostByTimePeriod(timePeriod, hourCount)
-  
-  // 更新顯示的即時費用
-  hourlyCostDisplay.value = hourlyCost
-}
-
-// 根據時段選擇計算鐘點制費用
-function calculateHourlyCostByTimePeriod(timePeriod, hourCount) {
-  // 根據時段選擇計算費用
-  switch (timePeriod) {
-    case '早班':
-      return hourCount * 200
-    case '晚班':
-      return hourCount * 250
-    case '夜班':
-      return hourCount * 300
-    default:
-      return 0
-  }
-}
-
-
-
-// 監聽夜間時段切換
-watch(isNightShift, (newVal) => {
-  toggleAdditionalService('HR09')
-})
-
-// 監聽加價急徵切換
-watch(isUrgent, (newVal) => {
-  toggleAdditionalService('HR10')
-})
-
-// 監聽二次加價急徵切換
-watch(isDoubleUrgent, (newVal) => {
-  toggleAdditionalService('HR11')
-})
-
-// 生命週期鎖子
-onMounted(() => {
- 
-  // 為每個項目添加自定義屬性，用於 GSAP 動畫
-  setTimeout(() => {
-    document.querySelectorAll('.service-item').forEach(item => {
-      const code = item.querySelector('.q-item-label')?.textContent?.trim()
-      if (code) {
-        item.setAttribute('data-code', code)
-      }
-    })
-    
-    // 初始化總費用元素的動畫
-    if (totalCostElement.value) {
-      gsap.fromTo(totalCostElement.value, 
-        { scale: 0.5, opacity: 0 },
-        { scale: 1, opacity: 1, duration: 1, ease: 'elastic.out(1, 0.5)' }
-      )
-    }
-    
-    // 為卡片添加入場動畫
-    const cards = document.querySelectorAll('.main-card, .summary-card')
-    cards.forEach((card, index) => {
-      gsap.fromTo(card, 
-        { y: 50, opacity: 0 },
-        { y: 0, opacity: 1, duration: 0.8, delay: 0.2 * index, ease: 'power2.out' }
-      )
-    })
-  }, 500)
-})
+const {
+  particlesLoaded,
+  careItems,
+  selectedHourlyItems,
+  selectedShiftItems,
+  selectedCategory,
+  selectedShiftType,
+  searchText,
+  showHelp,
+  previousTotalCost,
+  totalCostElement,
+  particleContainer,
+  tipAmount,
+  hourCount,
+  dayCount,
+  shiftDayCount,
+  showFilters,
+  showCalculator,
+  showShiftCalculator,
+  showAdditionalItems,
+  isNightShift,
+  isUrgent,
+  isDoubleUrgent,
+  activeFilters,
+  isAnyFilterActive,
+  particlesOptions,
+  selectedItems,
+  hourlyRate,
+  hourlyRateReachedLimit,
+  wouldExceedLimit,
+  hourlyTotalWithTime,
+  availableSubCategories,
+  shiftTypeItems,
+  filteredItems,
+  shiftTotalWithDays,
+  totalCost,
+  chartData,
+  chartOptions,
+  formatCurrency,
+  toggleItem,
+  createParticleExplosion,
+  isSelected,
+  selectShiftType,
+  resetSelections,
+  applyFilter,
+  resetAllFilters,
+  toggleAdditionalService,
+  calculatePreviewCost,
+  getItemIcon,
+  getItemColor
+} = useCareService()
 </script>
 
 <style scoped>
