+++ src/composables/useCareService.js
@@ -1,0 +1,24 @@
+import useCareState from './useCareState'
+import useCareFilters from './useCareFilters'
+import useCostCalculations from './useCostCalculations'
+import useParticleEffects from './useParticleEffects'
+import useCareActions from './useCareActions'
+import useChart from './useChart'
+
+export default function useCareService() {
+  const state = useCareState()
+  const filters = useCareFilters(state)
+  const calculations = useCostCalculations(state)
+  const particles = useParticleEffects(state, calculations)
+  const actions = useCareActions(state, calculations, particles)
+  const chart = useChart(calculations)
+
+  return {
+    ...state,
+    ...filters,
+    ...calculations,
+    ...particles,
+    ...actions,
+    ...chart
+  }
+}
